联合修改2：指令译码bug修复

+++ src/main/java/scau/os/soos/module/memory/MemoryController.java
@@ -34,7 +34,7 @@ public class MemoryController implements Module {
         return memoryService.read(address);
     }
 
-    public boolean write(int address, byte content) {
+    public boolean write(int address, int content) {
         return memoryService.write(address, content);
     }
